Nighthawk/requirements.txt
opencv_python==3.4.5.20
Flask==1.1.2
imageio==2.9.0
tensorflow==1.15.0
torch==1.5.1
Werkzeug==1.0.1
numpy==1.16.1
androguard==3.3.5
matplotlib==3.3.0
torchvision==0.6.1
Flask_SocketIO==5.0.1
networkx==2.5
frida==14.2.13
gym==0.18.0
Pillow==8.1.0
